Refactor: extracted more stuff from the main controller
